[common] Split `common_dep` into several dependencies
Currently, all `common/` source files are supplied via a single
`common_dep`. However in some cases, we may only want to reuse the part
of the utility functions in `common/` without importing the other part
of Gramine-supplied mini-libc (which might lead to colliding issues).

This commit separates `common/` into several dependencies for more
flexible usage.
