+++ common/src/path_utils.c
@@ -1,13 +1,23 @@
 /* SPDX-License-Identifier: LGPL-3.0-or-later */
-/* Copyright (C) 2014 Stony Brook University */
+/* Copyright (C) 2019 Invisible Things Lab
+ *                    Borys Popławski <borysp@invisiblethingslab.com>
+ */
 
 /*
  * This file contains functions to read app config (manifest) file and create a tree to
  * lookup / access config values.
  */
 
+#include <stddef.h>
+
+#ifdef USE_STDLIB
+#include <assert.h>
+#include <string.h>
+#else
 #include "api.h"
-#include "pal_error.h"
+#endif
+
+#include "path_utils.h"
 
 /*
  * Finds next '/' in `path`.
@@ -53,14 +63,14 @@ static inline bool find_prev_slash_offset(const char* path, size_t* offset) {
  * After returning it holds number of bytes actually written to it (including the ending '\0'). This
  * number is never greater than the size of the input path.
  */
-int get_norm_path(const char* path, char* buf, size_t* inout_size) {
+bool get_norm_path(const char* path, char* buf, size_t* inout_size) {
     assert(path && buf && inout_size);
     size_t path_size = strlen(path) + 1;
-    __UNUSED(path_size);  // used only for an assert at the end
+    (void)(path_size);  // used only for an assert at the end
 
     size_t size = *inout_size;
     if (!size) {
-        return -PAL_ERROR_INVAL;
+        return false;
     }
     /* reserve 1 byte for ending '\0' */
     size--;
@@ -73,7 +83,7 @@ int get_norm_path(const char* path, char* buf, size_t* inout_size) {
     /* handle an absolute path */
     if (is_absolute_path) {
         if (size < 1) {
-            return -PAL_ERROR_TOOLONG;
+            return false;
         }
         *buf++ = '/';
         size--;
@@ -93,7 +103,7 @@ int get_norm_path(const char* path, char* buf, size_t* inout_size) {
                 /* append undiscardable ".." since there is no previous token
                  * but only if the path is not absolute */
                 if (need_slash + 2u > size) {
-                    return -PAL_ERROR_TOOLONG;
+                    return false;
                 }
                 if (need_slash) {
                     *buf++ = '/';
@@ -113,7 +123,7 @@ int get_norm_path(const char* path, char* buf, size_t* inout_size) {
         } else {
             size_t len = (size_t)(end - path);
             if (need_slash + len > size - offset) {
-                return -PAL_ERROR_TOOLONG;
+                return false;
             }
             if (need_slash) {
                 buf[offset++] = '/';
@@ -133,7 +143,7 @@ int get_norm_path(const char* path, char* buf, size_t* inout_size) {
     *inout_size = ret_size + offset + 1;
     assert(*inout_size <= path_size);
 
-    return 0;
+    return true;
 }
 
 /*
@@ -141,9 +151,9 @@ int get_norm_path(const char* path, char* buf, size_t* inout_size) {
  * Before calling this function *size should hold the size of buf.
  * After returning it holds number of bytes actually written to it (including the trailing '\0').
  */
-int get_base_name(const char* path, char* buf, size_t* inout_size) {
+bool get_base_name(const char* path, char* buf, size_t* inout_size) {
     if (!path || !buf || !inout_size) {
-        return -PAL_ERROR_INVAL;
+        return false;
     }
 
     const char* end;
@@ -153,7 +163,7 @@ int get_base_name(const char* path, char* buf, size_t* inout_size) {
 
     size_t result = (size_t)(end - path);
     if (result + 1 > *inout_size) {
-        return -PAL_ERROR_TOOLONG;
+        return false;
     }
 
     memcpy(buf, path, result);
@@ -161,7 +171,7 @@ int get_base_name(const char* path, char* buf, size_t* inout_size) {
 
     *inout_size = result + 1;
 
-    return 0;
+    return true;
 }
 
 bool is_dot_or_dotdot(const char* name) {
