+++ common/src/string/atoi.c
@@ -43,23 +43,6 @@ static void begin_number(const char* str, int base, const char** out_s, int* out
     *out_sign = sign;
 }
 
-static int parse_digit(char c, int base) {
-    int digit;
-
-    if (c >= '0' && c <= '9') {
-        digit = c - '0';
-    } else if (c >= 'a' && c <= 'z') {
-        digit = c - 'a' + 10;
-    } else if (c >= 'A' && c <= 'Z') {
-        digit = c - 'A' + 10;
-    } else {
-        return -1;
-    }
-    if (digit >= base)
-        return -1;
-    return digit;
-}
-
 long strtol(const char* str, char** out_end, int base) {
     bool nothing_parsed = true;
     const char* s;
@@ -89,35 +72,6 @@ long strtol(const char* str, char** out_end, int base) {
     if (out_end)
         *out_end = (char*)(nothing_parsed ? str : s);
     return value;
-}
-
-int str_to_ulong(const char* str, unsigned int base, unsigned long* out_value,
-                 const char** out_end) {
-    if (base == 16 && str[0] == '0' && str[1] == 'x')
-        str += 2;
-
-    unsigned long value = 0;
-    const char* s = str;
-    while (*s != '\0') {
-        int digit = parse_digit(*s, base);
-        if (digit == -1)
-            break;
-
-        if (__builtin_mul_overflow(value, base, &value))
-            return -1;
-
-        if (__builtin_add_overflow(value, digit, &value))
-            return -1;
-
-        s++;
-    }
-
-    if (s == str)
-        return -1;
-
-    *out_value = value;
-    *out_end = s;
-    return 0;
 }
 
 #ifdef __LP64__
@@ -154,36 +108,4 @@ long int atol(const char* str) {
         s++;
     }
     return value;
-}
-
-int parse_size_str(const char* str, uint64_t* out_val) {
-    const char* endptr = NULL;
-    unsigned long size;
-    int ret = str_to_ulong(str, 10, &size, &endptr);
-    if (ret < 0)
-        return -1;
-
-    unsigned long unit = 1;
-    if (*endptr == 'G' || *endptr == 'g') {
-        unit = 1024 * 1024 * 1024;
-        endptr++;
-    } else if (*endptr == 'M' || *endptr == 'm') {
-        unit = 1024 * 1024;
-        endptr++;
-    } else if (*endptr == 'K' || *endptr == 'k') {
-        unit = 1024;
-        endptr++;
-    }
-
-    if (__builtin_mul_overflow(size, unit, &size))
-        return -1;
-
-    if (*endptr != '\0')
-        return -1; /* garbage found after the size string */
-
-    if (OVERFLOWS(__typeof__(*out_val), size))
-        return -1;
-
-    *out_val = size;
-    return 0;
 }
