+++ common/src/string/util.c
@@ -40,25 +40,3 @@ char* alloc_concat3(const char* a, size_t a_len, const char* b, size_t b_len,
     buf[a_len + b_len + c_len] = '\0';
     return buf;
 }
-
-bool strstartswith(const char* str, const char* prefix) {
-    size_t prefix_len = strlen(prefix);
-    size_t str_len = strnlen(str, prefix_len);
-
-    if (str_len < prefix_len) {
-        return false;
-    }
-
-    return !memcmp(str, prefix, prefix_len);
-}
-
-bool strendswith(const char* str, const char* suffix) {
-    size_t str_len = strlen(str);
-    size_t suffix_len = strlen(suffix);
-
-    if (str_len < suffix_len) {
-        return false;
-    }
-
-    return !memcmp(&str[str_len - suffix_len], suffix, suffix_len);
-}
