+++ libos/src/fs/libos_fs.c
@@ -16,6 +16,7 @@
 #include "libos_utils.h"
 #include "list.h"
 #include "pal.h"
+#include "path_utils.h"
 #include "toml.h"
 #include "toml_utils.h"
